Refactor `server` api file
