+++ features/admin.organizations.v1/api/root-organizations/system-settings/use-remote-log-publishing-configuration.ts
@@ -1,0 +1,61 @@
+/**
+ * Copyright (c) 2024, WSO2 LLC. (https://www.wso2.com).
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import { store } from "@wso2is/admin.core.v1";
+import useRequest, {
+    RequestConfigInterface,
+    RequestErrorInterface,
+    RequestResultInterface
+} from "@wso2is/admin.core.v1/hooks/use-request";
+import { HttpMethods } from "@wso2is/core/models";
+import {
+    RemoteLogPublishingConfigurationInterface
+} from "../../../models/root-organizations/system-settings/remote-log-publishing";
+
+/**
+ * Hook to get the remote log publishing configurations.
+ *
+ * @param shouldFetch - Should fetch the configurations.
+ * @returns remote log publishing configurations.
+ */
+const useRemoteLogPublishingConfiguration = <
+    Data = RemoteLogPublishingConfigurationInterface[],
+    Error = RequestErrorInterface
+>(
+        shouldFetch: boolean = true
+    ): RequestResultInterface<Data, Error> => {
+    const requestConfig: RequestConfigInterface = {
+        headers: {
+            "Content-Type": "application/json"
+        },
+        method: HttpMethods.GET,
+        url: store.getState().config.endpoints.remoteLogging
+    };
+
+    const { data, error, isValidating, mutate } = useRequest<Data, Error>(shouldFetch ? requestConfig : null);
+
+    return {
+        data,
+        error: error,
+        isLoading: !error && !data,
+        isValidating,
+        mutate: mutate
+    };
+};
+
+export default useRemoteLogPublishingConfiguration;
